Bug on SOD showing different prefecture / outages
Resolving an issue where if the default prefecture is changed and the application restarted, the outages shown do not map to the newly set prefecture.

The root cause of this issue has been identified to be the asynchronous manner of the outages_context.dart, where the retrieval of the default prefecture was not been awaited for.