Performance improvement for outage list items
Associating list items with specific keys

By having unique keys the framework will sync the entries of the list (where all the outages are displayed) with the corresponding item on the list.

It will also compare the UI diffs more efficiently. If the keys are not unique, and for example the content of the first page of the list scrolls off - flutter tries to match and repaint the layout relying on the first entry of the list, all the way down - which is not efficient at all.